Fix issue not bumping snapshot and timestamp
If the repository worker is rebooted it has not anymore the last task
settings in the memory.
During the execution of `automatic_version_bump` by kaprien the runner
is not calling the update.

Even calling the update, the `config.runner.update` doesn't call try to
load the `last_task_settings`.

Closes #37

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>