Fixed bug when all targets are not found
It was sending to the 'unpublished_meta' in Redis a empty string.
This will cause an issue parsing the tasks.

The protection was done in two levels

In the Method add_targets and remove_targets.
In the Method that handles the 'unpublished_meta'.

Additionally a better response was done for that ans partial work for
the issue https://github.com/kaprien/kaprien-rest-api/issues/26

Closes #58

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>