Add intermediate state for add/remove_targets
This adds to the add/remove_targets an intermediate state as
"PUBLISHING".
This states means that the Roles were edited and submitted to
"unpublised_metas" in the Redis.

The function _publish_meta_state will wait and send the state back to
the Backend Service, until it is published by the job `publish_targets_meta`
publishes it.

Closes #8

Improves https://github.com/kaprien/kaprien/issues/17

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>