Improve unit test coverage
In repository_service_tuf_worker/interfaces.py:
Made coverage ignore (# pragma: no cover) the interfaces'
methods that have exceptions in the abstract classes
as they will never be raised (a TypeError will be raised instead
if an abstract method is not implemented).
The exceptions could be substituted with the "pass" keyword in
IStorage(StorageBackendInterface) class, but to keep
compatibility with the securesystemslib/storage.py implementation
of StorageBackendInterface of the securesystemslib package, we
keep it as is.

In tests/unit/tuf_repository_service_worker/services/storage/test_local.py:
- Changed default stub value of version.
Stubbed tests for LocalStorage.get should not have default stub
for version equal to 1, because that's also the default value when
a ValueError exception is called.
- Added stubbing ValueError for max(versions), increasing coverage.

In repository_service_tuf_worker/repository.py there are still some
unit tests that can be added (see the miss in coverage).

Closes #9

Signed-off-by: Konstantinos Papadopoulos <konpap1996@yahoo.com>