Add LocalKeyVault online key config settings
Add configuration settings in LocalKeyVault about the online key that
will be used for "timestamp", "snapshot", "targets"  and "bin-n" roles.
That will allow us to load the online key instead of generating it
which is in line with what we want when we support more key vaults.

It's planned that if the user doesn't provide a path to the online key
as recommended we will generate an online key for him.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>