chore: improve performance updating Targets in DB
This improves the performance of updating Targets in the SQL DB.

Previously, we used to update the targets one by one in the DB and
this commit made it in bulk.

In publishing new targets, we used to persist in the hashed bins with
new targets to the backend storage, update the snapshot, and update
the timestamp followed by the targets one by one in the SQL DB.

The last step was improved doing the update of all published targets
at once.

Closes #176

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>