Skip refresh settings when starting a subtask
There is no point in refreshing the repository settings when we launch
a subtask from a main task that has already updated the repository
settings.
That's what happens when we launch a new "publish_targets" from the
"repository._send_publish_targets_task()" function.

By default this optional argument "refresh_settings" is false to
make sure we don't break any of the current components launching a
worker task.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>