Remove automatic job publish targets
As discussed in issue #175, we removed the automatic job that
publishes targets every minute.

The publish targets are done to persist in the RSTUF Worker backend
storage of the updated metadata with the new targets.

1. We will give full control over the failure to add/remove targets
to the user. They can decide to reprocess the entire flow (add/remove
targets), or they can trigger publish targets.
2. The automatic job publishes targets creates confused when and who
publishes the targets if the automatic job wins in the race condition
with add/remove targets.
3. For the feature load/import data it can slow down the process and
it would require an way to disable the automatic jobs during this
process, adding a lot of confusion.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>