Add sub-task monitoring in the add/remove task
When adding/removing targets, it triggers a new job to publish the
targets.

This subtask could fail for some reason, and we don't monitor it.

This commit adds the monitoring for this task and will give details
to the user once it fails the parent task (add/remove targets).

Closes #190

The monitoring is done in the `_update_task` as part of checks.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>