Add a way to handle exceptions LockNotOwnedError
It adds a better way to handle the LockNotOwnedError exception.

The LockNotOwnedError happens when the task exceeds the timeout,
and another task owns the lock.
If the task time out, the lock is released. If it doesn't finish
properly, it will raise (fail) the task. Otherwise, the ignores the
error because another task didn't lock it.
Closes #151

The lock is now an optional configurable setting during the container deployment.
Closes #22

The publish targets also return the task properly when finished.
Closes #204

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>