Accept "publish_targets" in add/remove targets (#203)
* Make plantuml execute in headless mode

When running locally I keep getting the same HeadlessException over and
over again.
As suggested by the plantuml documentation:
https://plantuml.com/en/faq#239d64f675c3e515 I run it in headless mode
and this seems to help.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Accept "publish_targets" in add/remove targets

There is a new field that will be introduced in the RSTUF add and remove
API calls named "publish_targets".
This flag can be used by users to tell whether to start a subtask
publishing the latest targets changes.
By default, this flag will be enabled.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Apply feedback

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

---------

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>