Refactor IKeyVault and LocalKeyVault "get" API
Change the purpose of "IKeyVault.get()" and instead of returning a "Key"
object make it to return a signer and because of that I renamed it to
"get_signer()". We can do that because we want to load the private
key inside the "get()" call anyway which is used for signing.

Also, I have refactored "get_singer()" to generate a signer for the
online key that will be shared for all roles except root.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>