Flexible maintainer CD Workflow
This PR introduces a change in the `cd.yml` and adds the
`publish_container.yml  Workflow.

We release our components versions using the tags `v*`
The process of release has two steps:

1) Run the functional tests (FT) using the component to be released
`v*` tag (from `main` branch) against the latest releases of the
components.

2) If the (FT) finishes successfully, the `cd.yml` releases the new
Docker Image GitHub Registry and adds it to our Releases page in Github.
(Python package in case of the CLI).

If step 1 fails, the release breaks the compatibility with the current
release.

The problem we are solving is how to release a new version when it breaks
the compatibility.

This PR adds the workflow dispatch option for releases.
The maintainers can now go to the Actions > CD > Run Workflow,
hoose the `v*` tag release, and specify the version of the components it
is compatible with, including dev (main branch) to run the FT tests.

Using the Workflow dispatcher process, the Maintainers should bump to a
version that shows it is breaking compatibility and make it clear on the
release page.

We have a (release process)[https://repository-service-tuf.readthedocs.io/en/latest/devel/release.html#release-process]
and we are not changing it, only adding the Workflow that allows those actions.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>