Make tests repository and app tests independent (#222)
Currently, repository.py and app.py use the worker and repository
settings setup in repository_service_tuf_worker/__init__.py
The way they are configured now (without a function/class directly
in the module) means that when for the first time anything is imported
from repository_service_tuf_worker/ it results in configuring the
worker and repository settings.
As a side effect, this means that even if something is imported from
repository_service_tuf_worker/ later this won't cause a new set of
worker and repository settings to be set up.

This becomes a problem for our tests where in the test_repository.py
and test_app.py modules we want to make sure that we have a new set
of settings.
That is important because the worker and repository settings define
the behavior of the repository.MetadataRepository object.

I fix that problem by just replacing the configuration of the settings
to be done when a function is called instead on import.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>