Simplify LocalKeyVault args and change configure
I simplified LocalKeyVault arguments by removing the "path" argument or
"LOCAL_KEYVAULT_PATH" setting because LocalKeyVault does not have
any "write" functionality any more and this attribute is not required
anymore.
Instead, I renamed "LOCAL_KEYVAULT_NAME" setting (or "key_name" arg)
to "LOCAL_KEYVAULT__KEY_PATH" (or "key_path" arg) to represent a better
that this should be a "path" to the online key and not just the key
file name.

Additionally, I changed the "IKeyVault.configure()" API call description
and changed the "LocalKeyVault.configure()" implementation to try to
load the online key and with this validate that everything with the
LocalKeyVault configuration is correct.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>