Patch LocalKeyVault.configure for the tests
I had to patch all tests from "tests_app.py" and "tuf_repository.py"
because all of them require a "MetadataRepository" instance which calls
"LocalKeyVault.configure()" at some point.
I didn't want to copy the online key to a specific folder and then
remove it, but instead I wanted to patch "LocalKeyVault.configure".
Given how many tests there are I decided that the best course of
action was to add a conftest.py and fixtures that could be reused.

I also found out that it's important to patch the
"LocalKeyVault.configure" correctly by using monkeypatch as otherwise
if we stub it then for all tests that are run after that the "configure"
API call will be stubbed for them as well.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>