Make LOCAL_KEYVAULT_KEY_PASSWORD required
As "SSlibSigner.from_priv_key_uri()" doesn't accepts a secrets_handler
that returns a password equal to "None", it requires that the password
is a string, we had to decide whether to use an empty string as a
default for a password or to mandate it.
Together with Kairo, we decided it's a safety hazard if we use "" as
a default password, so we decided to require passwords for the keys.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>