Remove repository._load
I am removing repository._load() as we are making a mistake in passing
self.storagebackend() as an argument to "Metadata.from_file()" where it
expects an object from the type securesystemslib.StorageBackendInterface
but our self._storagebackend() doesn't implement that interface.
Additionally, there are more checks made inside LocalStorage.get()
and this will allow us to be more flexible when we introduce other
storage services.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>