Update LocalStorage.get() to follow the interface & remove _load() (#237)
* Update LocalStorage.get() to follow the interface

Make sure that LocalStorage.get() and IStorage.get() return the same
type: Metadata[T].
Currently, the LocalStorage.get() behaves like a context manager and
returns "BufferedReader" without a good reason for that and doesn't
follow what the interface has defined as a return type - Metadata[T].

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Remove repository._load

I am removing repository._load() as we are making a mistake in passing
self.storagebackend() as an argument to "Metadata.from_file()" where it
expects an object from the type securesystemslib.StorageBackendInterface
but our self._storagebackend() doesn't implement that interface.
Additionally, there are more checks made inside LocalStorage.get()
and this will allow us to be more flexible when we introduce other
storage services.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Fix bug LocalStorage.get() bug

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

---------

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>