Remove IKeyVault write functionality and refactor IKeyVault.get() (#210)
* Fix typos

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Remove the write functionality of all keyvaults

We have decided that our keyvaults will be read-only and we no longer
need the "put" method.
Instead to load the online key (which is used for all roles except root)
we use environmental variables.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Refactor IKeyVault and LocalKeyVault "get" API

Change the purpose of "IKeyVault.get()" and instead of returning a "Key"
object make it to return a signer and because of that I renamed it to
"get_signer()". We can do that because we want to load the private
key inside the "get()" call anyway which is used for signing.

Also, I have refactored "get_singer()" to generate a signer for the
online key that will be shared for all roles except root.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Address comments

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Simplify tests with changes of get() and sign()

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Remove unnecessary "online" prefix

It's unnecessary given that we write about it in our official
documentation and during the new bootstrap ceremony we make it clear
that there is an online key reused for all roles except root.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Add public key information to LocalKeyVault.get()

By adding the public key information in the LocalKeyVault.get() API call
we can use a standard way to load securesystemslib Signers by using the
"from_priv_key_uri()" API call.
This will become a standard when we add new key vault implementations.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Remove unnecessary LocalKeyVault arguments

We have removed all usages of "_secrets_file" and the "DynaConf"
configuration inside "LocalKeyVault" and we can remove them.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Fix error on missing optional keyvault settings

Currently, we test and run RSTUF worker instances with provided values
for all three of optional LocalKeyVault optional arguments:
LOCAL_KEYVAULT_KEY_PATH, LOCAL_KEYVAULT_KEY_PASSWORD and
LOCAL_KEYVAULT_KEY_TYPE.

That's why to properly test the RSTUF worker we should remove their
values from tox.ini.
Doing so and making an instance reveals we need to add some checks
when getting expected kwargs in the repository.refresh_settings for
storage and key vault backends.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Simplify docker-compose and add an online key file

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Simplify LocalKeyVault args and change configure

I simplified LocalKeyVault arguments by removing the "path" argument or
"LOCAL_KEYVAULT_PATH" setting because LocalKeyVault does not have
any "write" functionality any more and this attribute is not required
anymore.
Instead, I renamed "LOCAL_KEYVAULT_NAME" setting (or "key_name" arg)
to "LOCAL_KEYVAULT__KEY_PATH" (or "key_path" arg) to represent a better
that this should be a "path" to the online key and not just the key
file name.

Additionally, I changed the "IKeyVault.configure()" API call description
and changed the "LocalKeyVault.configure()" implementation to try to
load the online key and with this validate that everything with the
LocalKeyVault configuration is correct.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Patch LocalKeyVault.configure for the tests

I had to patch all tests from "tests_app.py" and "tuf_repository.py"
because all of them require a "MetadataRepository" instance which calls
"LocalKeyVault.configure()" at some point.
I didn't want to copy the online key to a specific folder and then
remove it, but instead I wanted to patch "LocalKeyVault.configure".
Given how many tests there are I decided that the best course of
action was to add a conftest.py and fixtures that could be reused.

I also found out that it's important to patch the
"LocalKeyVault.configure" correctly by using monkeypatch as otherwise
if we stub it then for all tests that are run after that the "configure"
API call will be stubbed for them as well.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Fix lint

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Fix loading a docker container without settings

There is still a problem loading a docker container without the optional
LocalKeyVault settings.
I fix that by providing a default value for each setting.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Make LOCAL_KEYVAULT_KEY_PASSWORD required

As "SSlibSigner.from_priv_key_uri()" doesn't accepts a secrets_handler
that returns a password equal to "None", it requires that the password
is a string, we had to decide whether to use an empty string as a
default for a password or to mandate it.
Together with Kairo, we decided it's a safety hazard if we use "" as
a default password, so we decided to require passwords for the keys.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Add forgoten required setting

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Remove "KEY" prefix before LocalKeyVault settings

Remove redundant "KEY" prefix before LocalKeyVault settings.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Give default on optional arguments

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Address comments

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Document important info about the online key

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Support RSTUF workers with container secrets

To supports users who deployed RSTUF worker using container secrets for
their passwords we need to open the generated file with secrets and
make sure we remove the extra new lines.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Address Kairo's comments

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Fix linting errors

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Fix error when accessing keyids

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

---------

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>