Add a way to handle exceptions LockNotOwnedError (#251)
* Add a way to handle exceptions LockNotOwnedError

It adds a better way to handle the LockNotOwnedError exception.

The LockNotOwnedError happens when the task exceeds the timeout,
and another task owns the lock.
If the task time out, the lock is released. If it doesn't finish
properly, it will raise (fail) the task. Otherwise, the ignores the
error because another task didn't lock it.
Closes #151

The lock is now an optional configurable setting during the container deployment.
Closes #22

The publish targets also return the task properly when finished.
Closes #204

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Fix English/wording

Co-authored-by: Martin Vrachev <martin.vrachev@gmail.com>

* assert publish_targets datetime call

- Add assert for `publish_targets` `datetime.now()` call
- Move the mocking to end as it is called in the end of the function

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Improve locks

- Create a constant for the locks names
- On `publish_targets` uses lock to update timestamp (and snapshot)
- Remove unecessary bump for targets during `bump_online_roles`
- Updated tests

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Add comment to clarify the lock for timestamp

Add comment to clarify the lock for timestamp during the bump target
roles process

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

---------

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>
Co-authored-by: Martin Vrachev <martin.vrachev@gmail.com>