Repository Settings write implementation (#275)
* Repository Settings write implementation

This commit implements two new functionalities in the Repository

* `write_repository_settings`

This method has the functionality to write settings values to the
Repository settings. The repository settings uses Dynaconf configured to
the Redis Server (`RSTUF_REDIS_SERVER`). This configuration is shared
accross the Workers (if multiple).

This function receives a key and value and stores it to the Redis server

* `save_settings`

Uses the `write_repository_settings` to save all configuration from the
Repository, which is a defined payload from the RSTUF API service. this
function also requires the root TUF metadata.

settings payload:
```json
{
    "expiration": {
        "root": 365,
        "targets": 365,
        "snapshot": 1,
        "timestamp": 1,
        "bins": 1
    },
    "services": {
        "targets_base_url": "http://www.example.com/repository/",
        "number_of_delegated_bins": 4,
        "targets_online_key": True
    }
}
```

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Apply suggestions for the comments

Co-authored-by: Martin Vrachev <martin.vrachev@gmail.com>

* fix comment about giving threshold and num of keys

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

---------

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>
Co-authored-by: Martin Vrachev <martin.vrachev@gmail.com>