Add sub-feature: file as key body in base64
This feature supports using the key body directly in the
``RSTUF_LOCAL_KEYVAULT_KEYS``

It gives the possibility to avoid dealing with key files during the
deployment.
When a user deploys the RSTUF Worker previously, it requires the key
already available in the container volume (Persistent Volume or
Docker Volume for example).

Adding files to this volume is usually possible after a container
starts using tools such as ``docker cp`` or ``kubectl cp`` for
example. The container will fail because the file is missing during
the start.

This feature helps the user to give the file body in the
``RSTUF_LOCAL_KEYVAULT_KEYS``.

Example:
``RSTUF_LOCAL_KEYVAULT_KEYS=base64|YWI2ODU0YzE5,strongPass:online-rsa.key,strongPass,rsa``

The user can still use it inside a secret.

In practice, this problem:
https://github.com/repository-service-tuf/repository-service-tuf/blob/d6dca6203d78987f6742db706a47b49d2fdb84b1/.github/workflows/functional.yml#L127-L132

Closes #241

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>