Rafactored `metadata_rotation`:  `metadata_update`
It aligns the process name with the function name of the other
components (API and CLI).

Renames the function from `metadata_rotation` to `metadata_update`.

Add the `metadata_rotation` function, which calls the
`metadata_update` but also gives a warning and warning logs.
It gives compatibility for users using the RSTUF API v2.3.0b1 with
RSTUF Worker until version >= v0.4.0b1 > v1.0.0.

Fix the unit tests and add an extra test to check the deprecation warning.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>