AWS S3 Storage Backend Service implementation
Implements the AWS S3 Storage Backend Service

It uses the oficial `boto3` library for AWS Sessions, Client and
Resources an and also the `awswrangler` to query the AWS S3 metadata.

It implements all the unit tests covering 100% coverage.

Updates the RSTUF Worker documentation adding a better format for
Storage Backend and KeyVault Backend.

Closes #345

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>