Introduces the FT when PR is approved
This commit implements the Functional Tests when a PR is approved by a
reviewer.

It improves our checks on new changes in the RSTUF Worker, avoiding
breaking compatibility and later discovery during the release.

It covers any PR submitted, including the build PRs for updating the
dependencies.

It uses the Worker version from the PR source code after the merge
will also be used as `dev` tab.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>