Stop calling refresh_settings on each task
Currently, we are calling refresh_settings many times:
1. a couple of times when we set up SQL, Storage Backend, KeyVault
and timeout when initializing MetadataRepository
2. once per task we call refresh_settings

In a discussion with Kairo we reevaluated this design and realized
that if there is a change that needs to be made then the user needs
to update the environmental variables which will require restart
of all worker containers and as a result refresh_settings will be
called for sure. This means that us calling refresh_settings on
each task or during MetadataRepository initialization is redundant
as nothing can be changed from the settings without the requirement
to restart the containers, otherwise an error will be raised.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>