Use I(KeyVault|Storage).cofigure as factory method
Transform the I(KeyVault|Storage).cofigure as factory method meaning
it will be used as a constructor.
This will allow removing duplicating code from __init__ or configure.

Additionally, I used the opportunity to remove the "kwargs" variable
and operations inside interfaces._setup_service_dynaconf() as we
don't use it for anything. As a continuation, I was able also to remove
ServerSettings.argument field from ServerSettings as again it was not
used anywhere any more.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>