Simplify online_key addition to targets role
We don't need to access the keyvault to get the currently used signer
and from it to receive the public key dict, so that we can create
a new SSLibKey...
The reason we don't need to do that is because the public key we pass to
the keyvault is the same key we actually want to add to targets and is
from the correct type.
There is no need to go through all that work just to retreat a
SSLibKey.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>