In discussion with @kairoaraujo, we identified an opportunity to enhance the consistency of Task Results in our codebase.
Previously, the Celery Task result for `FAILURE` was structured as a
dictionary with keys `task_id`, `state` and `result`. The `result`
field was defined as either a `str` or `Dict[str, any]`, causing
inconsistency.

This commit introduces these changes to ensure a consistent data
structure and aligns with the specified API documentation.

For more details, refer to the discussions in this pull request
/pull/439